refactor: move error response record to exceptions package
