+++ back/src/main/java/mx/aplazo/bnpl/customers/CustomersController.java
@@ -8,8 +8,8 @@ import mx.aplazo.bnpl.customers.dto.request.CreateCustomerRequest;
 import mx.aplazo.bnpl.customers.dto.response.CustomerResponse;
 import mx.aplazo.bnpl.customers.exception.CustomerNotFoundException;
 import mx.aplazo.bnpl.exceptions.ErrorCode;
+import mx.aplazo.bnpl.exceptions.ErrorResponse;
 import mx.aplazo.bnpl.exceptions.ErrorResponseBuilder;
-import mx.aplazo.bnpl.validation.exception.ErrorResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
@@ -45,7 +45,8 @@ public class CustomersController {
   @ExceptionHandler(MethodArgumentNotValidException.class)
   public ResponseEntity<ErrorResponse> handleCustomerValidation(MethodArgumentNotValidException ex, HttpServletRequest request) {
     ErrorCode error = ErrorCode.INVALID_CUSTOMER_REQUEST;
-    return ErrorResponseBuilder.toResponse(ex.getMessage(), request, error);
+    String message = ex.getBindingResult().getFieldError().getDefaultMessage();
+    return ErrorResponseBuilder.toResponse(message, request, error);
   }
 
   @ExceptionHandler(CustomerNotFoundException.class)
