+++ back/src/main/java/mx/aplazo/bnpl/exceptions/ErrorResponse.java
@@ -1,4 +1,4 @@
-package mx.aplazo.bnpl.validation.exception;
+package mx.aplazo.bnpl.exceptions;
 
 public record ErrorResponse(
         String code,
