+++ back/src/main/java/mx/aplazo/bnpl/filters/RateLimitingFilter.java
@@ -5,8 +5,8 @@ import jakarta.servlet.*;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import mx.aplazo.bnpl.exceptions.ErrorCode;
+import mx.aplazo.bnpl.exceptions.ErrorResponse;
 import mx.aplazo.bnpl.exceptions.ErrorResponseBuilder;
-import mx.aplazo.bnpl.validation.exception.ErrorResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
